refactor(backend): domain reorg + router/service uniformity (v2)
Apply PR #30 (refactor/combined) intent to the current develop branch which
diverged significantly with new async article_fetcher, revision/lag/diff
endpoints, and additional services.

## AI / Models domain reorganisation

Move AI and model code out of the flat app/{ai,models}/ layer into
domain-scoped sub-packages:

- models/comparison/  – engine.py, models.py, registry.py, __init__.py
  (from ai/semantic_comparison, ai/model_registry, models/comparison*,
   models/section_comparison, models/response MissingInfo/ExtraInfo/
   ArticleComparisonResponse)
- models/extraction/  – engine.py, models.py, fact_extraction_models.json,
  __init__.py (from ai/fact_extraction, models/api_models Fact* classes)
- models/translation/ – engine.py, registry.py, models.py, __init__.py
  (from ai/translations, ai/translation_model_registry,
   models/comparison_request ChunkedTranslateRequest)
- models/wiki/        – structure.py, analysis.py, responses.py, __init__.py
  (from models/wiki_structure, models/structural_analysis,
   models/structured_response, models/response Source/TranslateArticle*)
- models/api.py       – ModelSelectionResponse, ListResponse
- models/server.py    – ServerModel (from models/server_model)

All original files converted to 1-line re-export shims for backward compat.

## Router / service uniformity

Add shared helpers to reduce duplication across routers:

- services/router_utils.py   – async resolve_and_fetch_article() wraps
  wiki_utils URL parsing + async article_fetcher; used by comparison router
- services/structured_translation.py – translate_article() extracted from
  structured_wiki router; used by structured-translated-article endpoint
- services/wiki_utils.py     – added sync helpers: parse_wikipedia_url(),
  resolve_title_and_lang(), validate_language_code(), _language_cache

## Consumer updates

All routers, services and tests updated to import from new domain paths:
- routers/comparison: duplicate chunked-text endpoint removed; imports fixed;
  asyncio.to_thread(article_fetcher) replaced with resolve_and_fetch_article
- routers/models, structural_analysis, wiki_articles: model import paths updated
- routers/structured_wiki: translate_article moved to service; import paths updated
- services/article_parser, citation/header/infobox/table_analysis,
  section_comparison: model import paths updated
- tests: patch targets updated to new module paths

All 189 tests pass.
