+++ symmetry-unified-backend/app/models/comparison/__init__.py
@@ -1,0 +1,92 @@
+from app.models.comparison.models import (
+    ArticleComparisonResponse,
+    BaseCompareRequest,
+    CompareRequest,
+    CompareResponse,
+    ComparisonResult,
+    ExtraInfo,
+    MissingInfo,
+    ParagraphDiff,
+    SectionCompareRequest,
+    SectionCompareResponse,
+    SectionDiff,
+    SemanticCompareRequest,
+    SentenceDiff,
+)
+from app.models.comparison.registry import COMPARISON_MODELS, DEFAULT_MODEL
+from app.models.comparison.engine import (
+    perform_semantic_comparison,
+    preprocess_input,
+    semantic_compare,
+    sentences_diff,
+    universal_sentences_split,
+)
+
+__all__ = [
+    "ArticleComparisonResponse",
+    "BaseCompareRequest",
+    "CompareRequest",
+    "CompareResponse",
+    "ComparisonResult",
+    "ExtraInfo",
+    "MissingInfo",
+    "ParagraphDiff",
+    "SectionCompareRequest",
+    "SectionCompareResponse",
+    "SectionDiff",
+    "SemanticCompareRequest",
+    "SentenceDiff",
+    "COMPARISON_MODELS",
+    "DEFAULT_MODEL",
+    "perform_semantic_comparison",
+    "preprocess_input",
+    "semantic_compare",
+    "sentences_diff",
+    "universal_sentences_split",
+]
+from app.models.comparison.models import (
+    ArticleComparisonResponse,
+    BaseCompareRequest,
+    CompareRequest,
+    CompareResponse,
+    ComparisonResult,
+    ExtraInfo,
+    MissingInfo,
+    ParagraphDiff,
+    SectionCompareRequest,
+    SectionCompareResponse,
+    SectionDiff,
+    SemanticCompareRequest,
+    SentenceDiff,
+)
+from app.models.comparison.registry import COMPARISON_MODELS, DEFAULT_MODEL
+from app.models.comparison.engine import (
+    perform_semantic_comparison,
+    preprocess_input,
+    semantic_compare,
+    sentences_diff,
+    universal_sentences_split,
+)
+
+__all__ = [
+    "ArticleComparisonResponse",
+    "BaseCompareRequest",
+    "CompareRequest",
+    "CompareResponse",
+    "ComparisonResult",
+    "ExtraInfo",
+    "MissingInfo",
+    "ParagraphDiff",
+    "SectionCompareRequest",
+    "SectionCompareResponse",
+    "SectionDiff",
+    "SemanticCompareRequest",
+    "SentenceDiff",
+    "COMPARISON_MODELS",
+    "DEFAULT_MODEL",
+    "perform_semantic_comparison",
+    "preprocess_input",
+    "semantic_compare",
+    "sentences_diff",
+    "universal_sentences_split",
+]
