+++ symmetry-unified-backend/app/models/extraction/__init__.py
@@ -1,0 +1,36 @@
+from app.models.extraction.models import FactExtractionRequest, FactExtractionResponse
+from app.models.extraction.engine import (
+    extract_facts,
+    get_available_models,
+    get_model_config,
+    model_exists_on_hf,
+    validate_model,
+)
+
+__all__ = [
+    "FactExtractionRequest",
+    "FactExtractionResponse",
+    "extract_facts",
+    "get_available_models",
+    "get_model_config",
+    "model_exists_on_hf",
+    "validate_model",
+]
+from app.models.extraction.models import FactExtractionRequest, FactExtractionResponse
+from app.models.extraction.engine import (
+    extract_facts,
+    get_available_models,
+    get_model_config,
+    model_exists_on_hf,
+    validate_model,
+)
+
+__all__ = [
+    "FactExtractionRequest",
+    "FactExtractionResponse",
+    "extract_facts",
+    "get_available_models",
+    "get_model_config",
+    "model_exists_on_hf",
+    "validate_model",
+]
