+++ symmetry-unified-backend/app/models/extraction/fact_extraction_models.json
@@ -1,0 +1,33 @@
+[
+  {
+    "id": "distilbart-cnn-12-6",
+    "name": "DistilBART CNN 12-6",
+    "provider": "huggingface",
+    "model_name": "sshleifer/distilbart-cnn-12-6",
+    "description": "Distilled BART model fine-tuned on CNN/DailyMail, good for extractive-style summarization on CPU",
+    "task": "summarization",
+    "prompt_style": "plain",
+    "use_chat_template": false
+  },
+  {
+    "id": "t5-small",
+    "name": "T5 Small",
+    "provider": "huggingface",
+    "model_name": "google-t5/t5-small",
+    "description": "Lightweight T5 model at ~240MB, requires instruction prefix. Good CPU performance for summarization tasks.",
+    "task": "summarization",
+    "prompt_style": "prefix",
+    "use_chat_template": false,
+    "prompt_prefix": "summarize: "
+  },
+  {
+    "id": "openrouter/free",
+    "name": "OpenRouter Free Model",
+    "provider": "openrouter",
+    "model_name": "openrouter/free",
+    "description": "Free LLM model via OpenRouter API. Uses the best available free model. Requires OPENROUTER_API_KEY environment variable.",
+    "task": "chat",
+    "prompt_style": "instruction",
+    "use_chat_template": true
+  }
+]
