+++ symmetry-unified-backend/app/models/response.py
@@ -1,30 +1,3 @@
-from pydantic import BaseModel, Field
-from typing import List, Optional
-
-
-class MissingInfo(BaseModel):
-    sentence: str = Field()
-    index: int
-    similarity_score: Optional[float] = None
-
-
-class ExtraInfo(BaseModel):
-    sentence: str = Field()
-    index: int
-    similarity_score: Optional[float] = None
-
-
-class SourceArticleResponse(BaseModel):
-    sourceArticle: str
-    articleLanguages: List[str]
-
-
-class TranslateArticleResponse(BaseModel):
-    translatedArticle: str
-
-
-class ArticleComparisonResponse(BaseModel):
-    missing_info: List[MissingInfo]
-    extra_info: List[ExtraInfo]
-    model_name: str = Field(default="sentence-transformers/LaBSE", description="Name of the model used for comparison")
-    similarity_threshold: float = Field(default=0.75, ge=0.0, le=1.0, description="Similarity threshold used for comparison")
+# Shim: contents moved to app.models.wiki.responses and app.models.comparison.models
+from app.models.wiki.responses import SourceArticleResponse, TranslateArticleResponse  # noqa: F401
+from app.models.comparison.models import MissingInfo, ExtraInfo, ArticleComparisonResponse  # noqa: F401
