+++ symmetry-unified-backend/app/models/translation/models.py
@@ -1,0 +1,7 @@
+from pydantic import BaseModel, Field
+
+
+class ChunkedTranslateRequest(BaseModel):
+    source_language: str = Field(..., min_length=2, max_length=10)
+    target_language: str = Field(..., min_length=2, max_length=10)
+    text: str = Field(..., min_length=1)
