+++ symmetry-unified-backend/app/models/wiki_structure.py
@@ -1,31 +1,2 @@
-from pydantic import BaseModel, Field
-from typing import List, Optional
-
-
-class Citation(BaseModel):
-    model_config = {"frozen": True}
-    label: str = Field()
-    url: Optional[str] = Field(default=None)
-
-
-class Reference(BaseModel):
-    model_config = {"frozen": True}
-    label: str = Field(max_length=10000)
-    id: Optional[str] = Field(default=None, max_length=300)
-    url: Optional[str] = Field(default=None, max_length=2000)
-
-
-class Section(BaseModel):
-    title: str = Field()
-    raw_content: str
-    clean_content: str
-    citations: Optional[List[Citation]] = None
-    citation_position: Optional[List[str]] = None
-
-
-class Article(BaseModel):
-    title: str = Field()
-    lang: str = Field()
-    source: str = Field()
-    sections: List[Section]
-    references: List[Reference]
+# Shim: contents moved to app.models.wiki.structure
+from app.models.wiki.structure import *  # noqa: F401, F403
