+++ symmetry-unified-backend/app/routers/structural_analysis.py
@@ -1,8 +1,7 @@
 from fastapi import APIRouter, HTTPException, Path
 from starlette import status
-from app.models import (
+from app.models.wiki.analysis import (
     FinalAnalysisResponse,
-    MultiLanguageScoreResponse,
     AnalysisResultsResponse,
 )
 from app.services import (
