+++ symmetry-unified-backend/app/routers/wiki_articles.py
@@ -9,7 +9,7 @@ import wikipediaapi
 import pycountry
 from fastapi import APIRouter, Query, HTTPException, Request
 
-from app.models import SourceArticleResponse
+from app.models.wiki.responses import SourceArticleResponse
 from app.services.cache import get_cached_article, set_cached_article
 
 router = APIRouter(prefix="/symmetry/v1/wiki", tags=["wiki"])
