+++ symmetry-unified-backend/app/services/header_analysis.py
@@ -1,7 +1,7 @@
 from bs4 import BeautifulSoup
 from fastapi import HTTPException
 from starlette import status
-from app.models import HeaderCount
+from app.models.wiki.analysis import HeaderCount
 from app.services.wiki_utils import page_exists
 import requests
