+++ symmetry-unified-backend/app/services/wiki_utils.py
@@ -1,7 +1,84 @@
 from datetime import datetime
-from typing import List, Optional
+from typing import Dict, List, Optional
+from urllib.parse import urlparse, unquote
 
 import httpx
+
+
+# ---------------------------------------------------------------------------
+# Sync URL / language helpers (used by routers and router_utils)
+# ---------------------------------------------------------------------------
+
+
+def parse_wikipedia_url(url: str) -> tuple[str, str]:
+    """Parse a Wikipedia URL and return (lang, title). Raises ValueError on bad input."""
+    parsed_url = urlparse(url)
+
+    if not parsed_url.netloc.endswith(".wikipedia.org"):
+        raise ValueError("Invalid domain - must be Wikipedia")
+
+    parts = parsed_url.netloc.split(".")
+    if len(parts) != 3:
+        raise ValueError("Invalid URL format")
+
+    lang = parts[0]
+    if not lang.isalpha() or len(lang) > 2:
+        raise ValueError("Invalid language code")
+
+    if not parsed_url.path.startswith("/wiki/"):
+        raise ValueError("Invalid article path")
+
+    title = parsed_url.path[6:].replace("_", " ")
+
+    if not title:
+        raise ValueError("Empty article title")
+
+    return lang, unquote(title)
+
+
+def resolve_title_and_lang(query: str, default_lang: str) -> tuple[str, str]:
+    """Extract Wikipedia title and language from a URL or plain title string."""
+    if "://" in query:
+        parsed = urlparse(query)
+        if parsed and parsed.path.startswith("/wiki/"):
+            lang = parsed.netloc.split(".")[0]
+            title = unquote(parsed.path.split("/wiki/")[-1].replace("_", " "))
+            return title, lang
+        raise ValueError(f"Invalid Wikipedia URL: {query}")
+    return query.replace("_", " "), default_lang
+
+
+# Language validation cache
+_language_cache: Dict[str, bool] = {}
+
+
+def validate_language_code(language_code: str) -> bool:
+    if language_code in _language_cache:
+        return _language_cache[language_code]
+
+    try:
+        import pycountry
+
+        VALID_LANGUAGE_CODES = {
+            lang.alpha_2
+            for lang in pycountry.languages
+            if hasattr(lang, "alpha_2") and lang.alpha_2
+        } | {
+            lang.alpha_3
+            for lang in pycountry.languages
+            if hasattr(lang, "alpha_3") and lang.alpha_3
+        }
+    except Exception:
+        VALID_LANGUAGE_CODES = set()
+
+    is_valid = language_code in VALID_LANGUAGE_CODES
+    _language_cache[language_code] = is_valid
+    return is_valid
+
+
+# ---------------------------------------------------------------------------
+# Async Wikipedia API helpers
+# ---------------------------------------------------------------------------
 
 
 async def page_exists(title: str, source_language: str = "en") -> bool:
