+++ symmetry-unified-backend/tests/test_fact_extraction_cache.py
@@ -3,7 +3,7 @@ import anyio
 import torch
 from types import SimpleNamespace
 
-from app.ai import fact_extraction
+from app.models.extraction import engine as fact_extraction
 
 
 class DummyTokenizer:
